+++ Example3Swift/Example3Swift/MainViewController.swift
@@ -20,19 +20,14 @@ class MainViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
+
         leftButton.target = self.revealViewController()
         leftButton.action = #selector(PBRevealViewController.revealLeftView)
         
         rightButton.target = self.revealViewController()
         rightButton.action = #selector(PBRevealViewController.revealRightView)
-        
-        if (image == nil) {
-            self.revealViewController()?.delegate = UIApplication.sharedApplication().delegate as! AppDelegate
-            self.revealViewController()?.leftViewRevealWidth = 200
-            
-            self.revealViewController()?.rightPresentViewHierarchically = true
-        }
-        else {
+
+        if (image != nil) {
             imageView.image = image
         }
     }
@@ -42,5 +37,53 @@ class MainViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     
+    @IBAction func replaceLeftView(sender: UIBarButtonItem) {
+        var controller: UIViewController?
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        let nc = revealViewController().leftViewController as! UINavigationController
+        
+        if nc.topViewController! is MenuTableViewController {
+            controller = storyboard.instantiateViewControllerWithIdentifier("RightViewController")
+            controller!.title = "Menu"
+        }
+        else {
+            controller = storyboard.instantiateViewControllerWithIdentifier("MenuTableViewController")
+        }
+        
+        let newNc = UINavigationController(rootViewController: controller!)
+        revealViewController().setLeftViewController(newNc, animated: true)
+    }
+    
+    @IBAction func resizeLeftView(sender: UIBarButtonItem) {
+        if revealViewController().leftViewRevealWidth < 200 {
+            revealViewController().leftViewRevealWidth = 200
+        }
+        else {
+            revealViewController().leftViewRevealWidth = 180
+        }
+    }
+    
+    @IBAction func replaceRightView(sender: UIBarButtonItem) {
+        var controller: UIViewController?
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)        
+        
+        if revealViewController().rightViewController is RightViewController2 {
+            controller = storyboard.instantiateViewControllerWithIdentifier("RightViewController")
+            revealViewController().setRightViewController(controller, animated: true)
+        }
+        else {
+            controller = storyboard.instantiateViewControllerWithIdentifier("RightViewController2")
+            revealViewController().setRightViewController(controller, animated: true)
+        }
+    }
+    
+    @IBAction func resizeRightView(sender: UIBarButtonItem) {
+        if revealViewController().rightViewRevealWidth < 180 {
+            revealViewController().rightViewRevealWidth = 180
+        }
+        else {
+            revealViewController().rightViewRevealWidth = 160
+        }
+    }
     
 }