+++ Example3Swift/Example3Swift/RightViewController2.swift
@@ -1,0 +1,23 @@
+//
+//  RightViewController2.swift
+//  Example3Swift
+//
+//  Created by Patrick BODET on 28/08/2016.
+//  Copyright © 2016 iDevelopper. All rights reserved.
+//
+
+import UIKit
+
+class RightViewController2: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+}